// src/richtexteditor/src/common/interface.ts
export interface IEditorModel {
    currentDocument?: Document
    execCommand?: Function
    observer?: Observer
    markdownSelection?: MarkdownSelection
    undoRedoManager?: UndoRedoManager | UndoRedoCommands
    nodeSelection?: NodeSelection
    mdSelectionFormats?: MDSelectionFormats
    domNode?: DOMNode
    nodeCutter?: NodeCutter
    formatPainterEditor?: IFormatPainterEditor
    domTree?: DOMMethods
    linkObj?: LinkCommand | MDLink
    videoObj?: VideoCommand
    autoFormatObj?: AutoFormatPlugin
    audioObj?: AudioCommand
    imgObj?: ImageCommand
    formatPainterObj?: FormatPainterActions
    tableObj?: TableCommand | MDTable
    pasteObj?: PasteCleanupAction
    emojiPickerObj?: EmojiPickerAction
    editableElement?: Element
    userAgentData?: CustomUserAgentData
    destroy?(): void
    beforeSlashMenuApplyFormat?(): void
    codeBlockObj?: CodeBlockPlugin
    isEntireRTEContentSelected?(): boolean
    aiAssistantActionObj?: AIAssistantActions
    isBlazor?: boolean
}